+++ cookbooks/arcgis-egdb/metadata.rb
@@ -1,13 +1,13 @@
-name 'arcgis-egdb'
-maintainer 'Esri'
+name             'arcgis-egdb'
+maintainer       'Esri'
 maintainer_email 'contracts@esri.com'
-license 'Apache 2.0'
-description 'Creates enterprise geodatabases in SQL Server or PostgreSQL DBMS and registers them with ArcGIS Server.'
+license          'Apache 2.0'
+description      'Creates enterprise geodatabases in SQL Server or PostgreSQL DBMS and registers them with ArcGIS Server.'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version '2.3.0'
-chef_version '>= 15.3' if defined? chef_version
+version          '2.4.0'
+chef_version     '>= 15.3', '< 19'
 
-depends    'arcgis-enterprise', '~> 5.3'
+depends    'arcgis-enterprise', '~> 5.4'
 
 supports   'windows'
 supports   'ubuntu'