+++ cookbooks/arcgis-pro/metadata.rb
@@ -4,10 +4,10 @@ maintainer_email 'contracts@esri.com'
 license          'Apache 2.0'
 description      'Installs and configures ArcGIS Pro'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '5.3.0'
-chef_version     '>= 15.3' if defined? chef_version
+version          '5.4.0'
+chef_version     '>= 15.3', '< 19'
 
-depends          'arcgis-repository', '~> 5.3'
+depends          'arcgis-repository', '~> 5.4'
 
 supports         'windows'