+++ cookbooks/arcgis-geoevent/CHANGELOG.md
@@ -3,6 +3,10 @@ arcgis-geoevent cookbook CHANGELOG
 
 This file is used to list changes made in each version of the arcgis-geoevent cookbook.
 
+5.4.0
+-----
+- Added support for ArcGIS GeoEvent Server 12.1.
+
 5.3.0
 -----
 - Added support for ArcGIS GeoEvent Server 12.0.