+++ cookbooks/arcgis-license-manager/CHANGELOG.md
@@ -3,6 +3,11 @@ arcgis-license-manager cookbook CHANGELOG
 
 This file is used to list changes made in each version of the arcgis-license-manager cookbook.
 
+5.4.0
+-----
+
+- Changed supported Chef versions.
+
 5.3.0
 -----