+++ cookbooks/arcgis-pro/CHANGELOG.md
@@ -3,6 +3,14 @@ arcgis-pro cookbook CHANGELOG
 
 This file is used to list changes made in each version of the arcgis-pro cookbook.
 
+5.4.0
+-----
+- Added support for ArcGIS Pro 3.7.
+
+5.3.0
+-----
+- Added support for ArcGIS Pro 3.6.
+
 5.2.0
 -----
 - Added support for ArcGIS Pro 3.5.